Integrated ImGui and changed the AddComponent method of GameObject
ImGui was compiled into a .lib and .dll and added to the project. Also added a scenegraph with imgui and an fps counter.

Add component was updated so that it is impossible to create a component without a game object who will automatically bind it to that game object. this was made so a component can never exist without an owner