Multiple Render Target support and display of hitboxes
Added the option to add multiple render targets to use as background/foreground.
The hitboxes can now be visualized by pressing the checking the checkbox in the settings.
The hitboxes now update its position every frame.
Added the ability to draw basic geometric shapes that can be used for debug rendering.
Fixed bugs for the ImGui Scenegraph